Complete Overhaul of Attack System
Complete overhaul of the attack system changed the following:
Each sprite of an animation can be set independently (including hitboxes, hurtboxes, and sprite duration)
Animations are controlled with a manual AnimationController component. All sprites should be attached to the AnimationController component and given labelled with the State's name and the duration to show each sprite.